requirements.txt
cycler==0.10.0
Cython==0.29.12
h5py==2.10.0
imageio==2.5.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.16.4
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
openslide-python==1.1.1
pandas==1.0.4
patsy==0.5.1
Pillow==6.1.0
pycocotools==2.0
pyparsing==2.4.2a1
python-dateutil==2.8.0
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.0
six==1.12.0
sklearn==0.0
statsmodels==0.11.1
threadpoolctl==2.1.0
torch==1.1.0
torchsummary==1.5.1
torchvision==0.3.0
tqdm==4.36.1
